+++ manufacturing-status/pom.xml
@@ -10,12 +10,22 @@
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>manufacturing-status</artifactId>
+
+    <modules>
+        <module>../public-module</module>
+    </modules>
     <dependencies>
         <!-- https://mvnrepository.com/artifact/org.springframework/spring-messaging -thisi is the version that state-machine 2.0.1 REALSE used-->
         <dependency>
             <groupId>org.springframework</groupId>
             <artifactId>spring-messaging</artifactId>
             <version>5.3.4</version>
+        </dependency>
+        <dependency>
+            <groupId>com.mytech</groupId>
+            <artifactId>public-module</artifactId>
+            <version>0.0.1-SNAPSHOT</version>
+            <scope>compile</scope>
         </dependency>
 
     </dependencies>
