--- manufacturing-status/src/main/java/repository/ManufacturingStatusRepositoryImpl.java
@@ -1,16 +1,0 @@
-package repository;
-
-import org.springframework.web.context.annotation.RequestScope;
-
-/**
- * @author `<a href="mailto:qiang.wang@1020@gmail.com">qiang</a>`
- * @date 2021-03-26
- * @description :
- */
-@RequestScope
-public class ManufacturingStatusRepositoryImpl {
-
-    void updateManufacturingStatus(String barcode) {
-
-    }
-}//ManufacturingStatusRepositoryImpl
