+++ pom.xml
@@ -9,13 +9,13 @@
         <module>key-parameter</module>
         <module>order-management</module>
         <module>manufacturing-status</module>
-        <module>new-module</module>
         <module>key-parameter</module>
+        <module>public-module</module>
     </modules>
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.5.3</version>
+        <version>2.4.3</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>com.mytech</groupId>
@@ -56,7 +56,13 @@
             <artifactId>guava</artifactId>
             <version>30.1-jre</version>
         </dependency>
+        <dependency>
+            <groupId>mysql</groupId>
+            <artifactId>mysql-connector-java</artifactId>
+            <scope>runtime</scope>
+        </dependency>
     </dependencies>
+
 
     <build>
         <plugins>
