+++ process-control/pom.xml
@@ -10,16 +10,35 @@
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>process-control</artifactId>
+
     <!-- 这一部分必须手工添加-->
     <modules>
         <module>../manufacturing-status</module>
+        <!-- <module>../public-module</module> -->
     </modules>
+
+    <properties>
+        <spirng-statemachine.version>2.0.1.RELEASE</spirng-statemachine.version>
+    </properties>
+
     <dependencies>
         <dependency>
             <groupId>com.mytech</groupId>
             <artifactId>manufacturing-status</artifactId>
             <version>0.0.1-SNAPSHOT</version>
             <scope>compile</scope>
+        </dependency>
+        <dependency>
+            <groupId>com.mytech</groupId>
+            <artifactId>public-module</artifactId>
+            <version>0.0.1-SNAPSHOT</version>
+            <scope>compile</scope>
+        </dependency>
+        <!-- https://mvnrepository.com/artifact/org.springframework.statemachine/spring-statemachine-core -->
+        <dependency>
+            <groupId>org.springframework.statemachine</groupId>
+            <artifactId>spring-statemachine-core</artifactId>
+            <version>${spirng-statemachine.version}</version>
         </dependency>
     </dependencies>
