+++ product-result/src/main/java/com/mytech/dto/ProductionResultOutputDTO.java
@@ -1,0 +1,23 @@
+package com.mytech.dto;
+
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.ToString;
+import lombok.experimental.SuperBuilder;
+import lombok.experimental.Tolerate;
+
+/**
+ * @author `<a href="mailto:qiang.wang@1020@gmail.com">qiang</a>`
+ * @date 2021-01-11
+ * @description :
+ */
+@Data
+@SuperBuilder
+@ToString(callSuper = true)
+@EqualsAndHashCode(callSuper = true)
+public class ProductionResultOutputDTO extends ProductionResultDTO {
+
+    @Tolerate
+    public ProductionResultOutputDTO(){}
+
+}
