--- product-result/src/main/java/dto/ProductionResultInputDTO.java
@@ -1,21 +1,0 @@
-package src.main.java.dto;
-
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.ToString;
-import lombok.experimental.SuperBuilder;
-import lombok.experimental.Tolerate;
-
-/**
- * @author `<a href="mailto:qiang.wang@1020@gmail.com">qiang</a>`
- * @date 2021-01-11
- * @description :
- */
-@Data
-@ToString(callSuper = true)
-@EqualsAndHashCode(callSuper = true)
-@SuperBuilder
-public class ProductionResultInputDTO extends ProductionResultDTO {
-    @Tolerate
-    public ProductionResultInputDTO(){}
-}
