+++ public-module/pom.xml
@@ -9,21 +9,18 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>product-result</artifactId>
-    <modules>
-        <!---public module is only to use the application.yaml-->
-        <module>../public-module</module>
-    </modules>
+    <artifactId>public-module</artifactId>
 
     <dependencies>
-        <!---add public-module dependency is only to use the application.yaml-->
-        <dependency>
-            <groupId>com.mytech</groupId>
-            <artifactId>public-module</artifactId>
-            <version>0.0.1-SNAPSHOT</version>
-            <scope>compile</scope>
-        </dependency>
-    </dependencies>
+        <!-- https://mvnrepository.com/artifact/org.springframework.boot/spring-boot-starter-actuator -->
 
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-actuator</artifactId>
+            <version>2.4.3</version>
+        </dependency>
+
+
+    </dependencies>
 
 </project>
